# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": ["code", "message", "details"],
        "title": "ApiError",
        "type": "object"
      },
...
      "CompetencyQuestionModel": {
        "additionalProperties": false,
        "properties": {
          "acceptance_criterion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Criterion"
          },
          "expected_boolean": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Boolean"
          },
          "iri": {
            "title": "Iri",
            "type": "string"
          },
          "minimum_result_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Result Count"
          },
          "module": {
            "title": "Module",
            "type": "string"
          },
          "query_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query File"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "iri",
          "text",
          "module",
          "state",
          "query_file",
          "expected_boolean",
          "minimum_result_count",
          "acceptance_criterion"
        ],
        "title": "CompetencyQuestionModel",
        "type": "object"
      },
      "CompetencyQuestionPageModel": {
        "additionalProperties": false,
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompetencyQuestionModel"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "offset", "limit", "has_next"],
        "title": "CompetencyQuestionPageModel",
        "type": "object"
      },
      "CompetencyResultModel": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "$ref": "#/components/schemas/CompetencyQuestionModel"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SparqlResultModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": ["passed", "failed", "not_executable"],
            "title": "Status",
            "type": "string"
          }
        },
        "required": ["question", "status", "reason", "result"],
        "title": "CompetencyResultModel",
        "type": "object"
      },
      "CompetencyRunRequest": {
        "additionalProperties": false,
        "properties": {
          "iri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iri"
          }
        },
        "title": "CompetencyRunRequest",
        "type": "object"
      },
...
      "ImpactModel": {
        "additionalProperties": false,
        "properties": {
          "affected_importers": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Affected Importers",
            "type": "array"
          },
          "ancestors": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Ancestors",
            "type": "array"
          },
          "competency_questions": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Competency Questions",
            "type": "array"
          },
          "descendants": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Descendants",
            "type": "array"
          },
          "import_dependencies": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Import Dependencies",
            "type": "array"
          },
          "incoming": {
            "items": {
              "$ref": "#/components/schemas/QuadModel"
            },
            "title": "Incoming",
            "type": "array"
          },
          "outgoing": {
            "items": {
              "$ref": "#/components/schemas/QuadModel"
            },
            "title": "Outgoing",
            "type": "array"
          },
          "predicate_uses": {
            "items": {
              "$ref": "#/components/schemas/QuadModel"
            },
            "title": "Predicate Uses",
            "type": "array"
          },
          "resource": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "shapes": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Shapes",
            "type": "array"
          }
        },
        "required": [
          "resource",
          "incoming",
          "outgoing",
          "predicate_uses",
          "ancestors",
          "descendants",
          "shapes",
          "competency_questions",
          "import_dependencies",
          "affected_importers"
        ],
        "title": "ImpactModel",
        "type": "object"
      },
...
      "QuadModel": {
        "additionalProperties": false,
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "object": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "predicate": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "priority": {
            "default": 7,
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          },
          "relationship_kind": {
            "default": "other",
            "enum": [
              "subclass",
              "subproperty",
              "domain",
              "range",
              "import",
              "internal_object_property",
              "other"
            ],
            "title": "Relationship Kind",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "subject": {
            "$ref": "#/components/schemas/RdfValueModel"
          }
        },
        "required": ["subject", "predicate", "object", "graph"],
        "title": "QuadModel",
        "type": "object"
      },
      "RdfValueModel": {
        "additionalProperties": false,
        "properties": {
          "compact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compact"
          },
          "datatype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datatype"
          },
          "kind": {
            "enum": ["iri", "bnode", "literal"],
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": ["kind", "value"],
        "title": "RdfValueModel",
        "type": "object"
      },
...
      "SparqlRequest": {
        "additionalProperties": false,
        "properties": {
          "max_results": {
            "default": 1000,
            "maximum": 5000.0,
            "minimum": 1.0,
            "title": "Max Results",
            "type": "integer"
          },
          "query": {
            "maxLength": 65536,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "timeout_seconds": {
            "default": 5.0,
            "exclusiveMinimum": 0.0,
            "maximum": 10.0,
            "title": "Timeout Seconds",
            "type": "number"
          }
        },
        "required": ["query"],
        "title": "SparqlRequest",
        "type": "object"
      },
      "SparqlResultModel": {
        "additionalProperties": false,
        "properties": {
          "boolean": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boolean"
          },
          "kind": {
            "enum": ["select", "ask", "construct", "describe"],
            "title": "Kind",
            "type": "string"
          },
          "rows": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RdfValueModel"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Rows",
            "type": "array"
          },
          "triples": {
            "items": {
              "items": {
                "$ref": "#/components/schemas/RdfValueModel"
              },
              "type": "array"
            },
            "title": "Triples",
            "type": "array"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          },
          "variables": {
            "items": {
              "type": "string"
            },
            "title": "Variables",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "variables",
          "rows",
          "boolean",
          "triples",
          "truncated"
        ],
        "title": "SparqlResultModel",
        "type": "object"
      },
...
      "ValidationIssueModel": {
        "additionalProperties": false,
        "properties": {
          "details": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Details",
            "type": "object"
          },
          "graph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "resource": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource"
          },
          "resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Type"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "rule_id",
          "severity",
          "message",
          "resource",
          "resource_type",
          "path",
          "graph",
          "details"
        ],
        "title": "ValidationIssueModel",
        "type": "object"
      },
      "ValidationReportModel": {
        "additionalProperties": false,
        "properties": {
          "conforms": {
            "title": "Conforms",
            "type": "boolean"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueModel"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": ["conforms", "counts", "issues"],
        "title": "ValidationReportModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Stable API over the canonical RDF/Git workspace, including controlled proposal authoring, semantic review and the synchronized canonical agent contract.",
    "title": "Enterprise Ontology Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/competency_questions": {
      "get": {
        "operationId": "list_competency_questions_api_competency_questions_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "module",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyQuestionPageModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "List Competency Questions",
        "tags": ["semantic"]
      }
    },
    "/api/competency_questions/run": {
      "post": {
        "operationId": "run_competency_questions_api_competency_questions_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompetencyResultModel"
                  },
                  "title": "Response Run Competency Questions Api Competency Questions Run Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Run Competency Questions",
        "tags": ["semantic"]
      }
    },
...
    "/api/impact": {
      "get": {
        "operationId": "impact_api_impact_get",
        "parameters": [
          {
            "in": "query",
            "name": "iri",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Iri",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Impact",
        "tags": ["semantic"]
      }
    },
...
    "/api/sparql/query": {
      "post": {
        "operationId": "sparql_query_api_sparql_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SparqlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SparqlResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Sparql Query",
        "tags": ["semantic"]
      }
    },
...
    "/api/validation/run": {
      "post": {
        "operationId": "run_validation_api_validation_run_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Run Validation",
        "tags": ["semantic"]
      }
    },
...
  }
}